backend/requirements.txt
pipecat-ai[elevenlabs,cartesia,openai,silero,deepgram]
fastapi
uvicorn
python-dotenv
loguru
twilio
langchain
langchain-openai
langchain-community
langgraph
langsmith
pymongo 
cryptography  
redis  
pydantic  
httpx  
pandas  
numpy  
bcrypt  
python-jose[cryptography]  
passlib 
prometheus-client  
structlog  